+++ Example/AftermathDemo/AftermathDemo/Sources/Configuration/AftermathConfigurator.swift
@@ -5,7 +5,5 @@ public struct AftermathConfigurator: Configurator {
   public func configure() {
     Engine.sharedInstance.use(PostsStory.Handler())
     Engine.sharedInstance.use(UsersStory.Handler())
-
-    
   }
 }
